Update README.md
How do I approach designing software?
Some of the skills I learned in this project was to ensure that everything was commented. I had to keep changing one line of code and it would either make or break my image, program, or even the texture. Staying organized made a huge difference in this.
Doing this kept me productive on my project. The biggest part was where I labeled which side and section each vertex was on. This kept it easy to move an entire object, a single vertex, and even the texture.
I will use this to keep all my future programs organized. Thanks to the way I made my program, it is very modular and doing it this way will allow me to be able to reuse some parts or even just change small sections of my code to come out with completely different images. 
How do I approach developing programs?
The biggest development strategy I used for this project was a simple one. I ended up going out and buying graph paper. This allowed me to have a physical copy of what my program would look like. On top of that. Because it was a graph paper, it made it super easy to find out where my points should go on my program. 
The iteration became a huge time consuming this. Especially when trying to make a simple object like a cube. You needed two lines of code per side. However, by staying organized it just became a copy and paste thing where you could just edit two points. This part of the iteration made it run fast after I figured it out. 
Each milestone was just the same program with another thing added to it. This made the final project not seem like such a big deal. Since it was all small sections, it made it much easier with way less anxiety when working on this project. 
How can computer science help me in reaching my goals?
This class gave me a much more in dept understanding about the time and effort some videogames go through. I have taken several years of Autodesk and Blender classes, and this seems like an extremely easier way to make some of the objects. However, OpenGL has some advantages over things like Autodesk so it’s important to keep that in mind.
I will take the information I have learned in this class and continue to work on it. I do love playing video games and if I can slowly work on a personal project that would be a great plus. However, I want to go into cyber security, and I do not think this will help me with my career path at all. I will keep playing around with it as a hobby though.